requirements.txt
requests~=2.31.0
lxml
PyExecJS
beautifulsoup4
urllib2_file
pycryptodome